[auth-map] Move ServiceBlock into AbstractServiceBlock
