--- src/main/scala/temple/ast/ServiceBlock.scala
@@ -1,10 +1,0 @@
-package temple.ast
-
-import temple.ast.Metadata.ServiceMetadata
-
-/** A service block, representing one microservice on its own isolated server */
-case class ServiceBlock(
-  attributes: Map[String, Attribute],
-  metadata: Seq[ServiceMetadata] = Nil,
-  structs: Map[String, StructBlock] = Map.empty,
-) extends GeneratedBlock
