+++ src/test/scala/temple/detail/LanguageSpecificDetailBuilderTestData.scala
@@ -1,8 +1,9 @@
 package temple.detail
 
+import temple.ast.AbstractServiceBlock._
 import temple.ast.AttributeType._
 import temple.ast.Metadata.ServiceLanguage
-import temple.ast.{Attribute, ProjectBlock, ServiceBlock, Templefile}
+import temple.ast.{Attribute, ProjectBlock, Templefile}
 
 import scala.collection.immutable.ListMap
